ref: 클린 아키텍처 패키지 구조를 잡았어요.
