+++ src/main/kotlin/org/teamalilm/alilmbe/domain/member/repository/MemberRepository.kt
@@ -1,7 +1,7 @@
 package org.teamalilm.alilmbe.domain.member.repository
 
 import org.springframework.data.jpa.repository.JpaRepository
-import org.teamalilm.alilmbe.adapter.out.persistence.jpa.entity.member.Member
+import org.teamalilm.alilmbe.adapter.out.persistence.entity.member.Member
 
 interface MemberRepository : JpaRepository<Member, Long> {
