+++ src/main/kotlin/org/teamalilm/alilmbe/global/security/service/oAuth2/handler/CustomSuccessHandler.kt
@@ -9,8 +9,8 @@ import org.springframework.security.web.authentication.SimpleUrlAuthenticationSu
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
 import org.springframework.web.util.UriComponentsBuilder
-import org.teamalilm.alilmbe.adapter.out.persistence.jpa.entity.member.Member
-import org.teamalilm.alilmbe.adapter.out.persistence.jpa.entity.member.Role
+import org.teamalilm.alilmbe.adapter.out.persistence.entity.member.Member
+import org.teamalilm.alilmbe.adapter.out.persistence.entity.member.Role
 import org.teamalilm.alilmbe.domain.member.repository.MemberRepository
 import org.teamalilm.alilmbe.global.security.jwt.JwtUtil
 import org.teamalilm.alilmbe.global.security.service.oAuth2.data.Provider
